moves test folder within sandx
